Add files via upload
Colombo District DSD Map

This project involves creating a Divisional Secretariat Division (DSD) map of the Colombo District using QGIS, as part of an exercise for my WebGIS module. The map was generated in QGIS and then exported to the web using the QGIS2Web plugin. This allowed me to create an interactive web map, which I have uploaded to GitHub for easy access and sharing.